Create Tree from inorder and postorder traversals
Root is going to be the last node in the postorder traversal. Starting from there, we will follow the pattern root, right, left. Hence keep track of this last index using pIndex. Once root is found in the post order traversal, all elements towards left of it in inorder will be a part of the left subtree and all elements towards right of it will be right subtree.